+++ clients/admin-ui/cypress/support/stubs.ts
@@ -271,3 +271,18 @@ export const stubTranslationConfig = (enabled: boolean) => {
     },
   }).as("getTranslationConfig");
 };
+
+export const stubStagedResourceActions = () => {
+  cy.intercept("POST", "/api/v1/plus/discovery-monitor/**/confirm*").as(
+    "confirmResource"
+  );
+  cy.intercept("POST", "/api/v1/plus/discovery-monitor/mute*").as(
+    "ignoreResource"
+  );
+  cy.intercept("POST", "/api/v1/plus/discovery-monitor/promote*").as(
+    "promoteResource"
+  );
+  cy.intercept("POST", "/api/v1/plus/discovery-monitor/un-mute*").as(
+    "unmuteResource"
+  );
+};