+++ clients/admin-ui/src/features/data-discovery-and-detection/ActionButton.tsx
@@ -1,0 +1,31 @@
+import { Button, ButtonProps, Text } from "@fidesui/react";
+import { ReactElement } from "react";
+
+const ActionButton = ({
+  title,
+  icon,
+  onClick,
+  disabled,
+  variant = "outline",
+  colorScheme = undefined,
+}: {
+  title: string;
+  icon: ReactElement;
+  onClick: () => void;
+  disabled?: boolean;
+} & Pick<ButtonProps, "variant" | "colorScheme">) => (
+  <Button
+    size="xs"
+    onClick={onClick}
+    disabled={disabled}
+    variant={variant}
+    colorScheme={colorScheme}
+    data-testid={`action-${title}`}
+  >
+    {icon}
+    <Text marginLeft={1} fontWeight="semibold" fontSize={12}>
+      {title}
+    </Text>
+  </Button>
+);
+export default ActionButton;