+++ clients/admin-ui/src/features/data-discovery-and-detection/hooks/useDetectionResultColumns.tsx
@@ -1,0 +1,128 @@
+import { ColumnDef, createColumnHelper } from "@tanstack/react-table";
+
+import { DefaultCell, DefaultHeaderCell } from "~/features/common/table/v2";
+import { RelativeTimestampCell } from "~/features/common/table/v2/cells";
+import DetectionItemAction from "~/features/data-discovery-and-detection/DetectionItemActions";
+import ResultStatusCell from "~/features/data-discovery-and-detection/tables/ResultStatusCell";
+import { DiscoveryMonitorItem } from "~/features/data-discovery-and-detection/types/DiscoveryMonitorItem";
+import { StagedResourceType } from "~/features/data-discovery-and-detection/types/StagedResourceType";
+import findResourceChangeType from "~/features/data-discovery-and-detection/utils/findResourceChangeType";
+
+import findProjectFromUrn from "../utils/findProjectFromUrn";
+
+const useDetectionResultColumns = ({
+  resourceType,
+}: {
+  resourceType?: StagedResourceType;
+}) => {
+  const columnHelper = createColumnHelper<DiscoveryMonitorItem>();
+
+  const defaultColumns: ColumnDef<DiscoveryMonitorItem, any>[] = [];
+
+  if (!resourceType) {
+    return { columns: defaultColumns };
+  }
+
+  if (resourceType === StagedResourceType.SCHEMA) {
+    const columns = [
+      columnHelper.accessor((row) => row.name, {
+        id: "name",
+        cell: (props) => <ResultStatusCell result={props.row.original} />,
+        header: (props) => <DefaultHeaderCell value="Name" {...props} />,
+      }),
+      columnHelper.accessor((row) => row.urn, {
+        id: "project",
+        cell: (props) => (
+          <DefaultCell value={findProjectFromUrn(props.getValue())} />
+        ),
+        header: (props) => <DefaultHeaderCell value="Project" {...props} />,
+      }),
+      columnHelper.display({
+        id: "type",
+        cell: () => <DefaultCell value="Dataset" />,
+        header: "Type",
+      }),
+      columnHelper.accessor((row) => row.monitor_config_id, {
+        id: "monitor",
+        cell: (props) => <DefaultCell value={props.getValue()} />,
+        header: (props) => <DefaultHeaderCell value="Detected by" {...props} />,
+      }),
+      columnHelper.accessor((row) => row.source_modified, {
+        id: "time",
+        cell: (props) => <RelativeTimestampCell time={props.getValue()} />,
+        header: (props) => <DefaultHeaderCell value="When" {...props} />,
+      }),
+      columnHelper.display({
+        id: "actions",
+        cell: (props) => <DetectionItemAction resource={props.row.original} />,
+        header: "Actions",
+      }),
+    ];
+    return { columns };
+  }
+
+  if (resourceType === StagedResourceType.TABLE) {
+    const columns = [
+      columnHelper.accessor((row) => row.name, {
+        id: "name",
+        cell: (props) => <ResultStatusCell result={props.row.original} />,
+        header: (props) => <DefaultHeaderCell value="Table name" {...props} />,
+      }),
+      columnHelper.accessor((row) => findResourceChangeType(row), {
+        id: "type",
+        cell: (props) => <DefaultCell value={props.getValue()} />,
+        header: (props) => <DefaultHeaderCell value="Type" {...props} />,
+      }),
+      columnHelper.accessor((row) => row.monitor_config_id, {
+        id: "monitor",
+        cell: (props) => <DefaultCell value={props.getValue()} />,
+        header: (props) => <DefaultHeaderCell value="Detected by" {...props} />,
+      }),
+      columnHelper.accessor((row) => row.source_modified, {
+        id: "time",
+        cell: (props) => <RelativeTimestampCell time={props.getValue()} />,
+        header: (props) => <DefaultHeaderCell value="When" {...props} />,
+      }),
+      columnHelper.display({
+        id: "actions",
+        cell: (props) => <DetectionItemAction resource={props.row.original} />,
+        header: "Actions",
+      }),
+    ];
+    return { columns };
+  }
+
+  if (resourceType === StagedResourceType.FIELD) {
+    const columns = [
+      columnHelper.accessor((row) => row.name, {
+        id: "name",
+        cell: (props) => <ResultStatusCell result={props.row.original} />,
+        header: (props) => <DefaultHeaderCell value="Field name" {...props} />,
+      }),
+      columnHelper.accessor((row) => findResourceChangeType(row), {
+        id: "type",
+        cell: (props) => <DefaultCell value={props.getValue()} />,
+        header: (props) => <DefaultHeaderCell value="Type" {...props} />,
+      }),
+      columnHelper.accessor((row) => row.monitor_config_id, {
+        id: "monitor",
+        cell: (props) => <DefaultCell value={props.getValue()} />,
+        header: (props) => <DefaultHeaderCell value="Detected by" {...props} />,
+      }),
+      columnHelper.accessor((row) => row.source_modified, {
+        id: "time",
+        cell: (props) => <RelativeTimestampCell time={props.getValue()} />,
+        header: (props) => <DefaultHeaderCell value="When" {...props} />,
+      }),
+      columnHelper.display({
+        id: "actions",
+        cell: (props) => <DetectionItemAction resource={props.row.original} />,
+        header: "Actions",
+      }),
+    ];
+    return { columns };
+  }
+  return { columns: defaultColumns };
+};
+
+export default useDetectionResultColumns;