+++ clients/admin-ui/src/features/data-discovery-and-detection/types/StagedResourceType.ts
@@ -1,0 +1,7 @@
+export enum StagedResourceType {
+  DATABASE = "Database",
+  SCHEMA = "Schema",
+  TABLE = "Table",
+  FIELD = "Field",
+  NONE = "none",
+}