+++ clients/admin-ui/src/types/api/models/BigQueryDocsSchema.ts
@@ -16,4 +16,8 @@ export type BigQueryDocsSchema = {
    * The dataset within your BigQuery project that contains the tables you want to access.
    */
   dataset: string;
+  /**
+   * The projects within your BigQuery organization that you want to scope your connection to.
+   */
+  project: Array<string>;
 };