+++ clients/admin-ui/src/types/api/models/Page_StagedResource_.ts
@@ -1,0 +1,13 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+import type { StagedResource } from "./StagedResource";
+
+export type Page_StagedResource_ = {
+  items: Array<StagedResource>;
+  total: number;
+  page: number;
+  size: number;
+  pages?: number;
+};