+++ clients/privacy-center/__tests__/common/geolocation.test.ts
@@ -34,6 +34,20 @@ describe("getGeolocation", () => {
       });
     });
 
+    it("supports the special 'EEA' code in country header", async () => {
+      const req = createRequest({
+        url: "https://privacy.example.com/fides.js",
+        headers: {
+          "CloudFront-Viewer-Country": "EEA",
+        },
+      });
+      const geolocation = await lookupGeolocation(req);
+      expect(geolocation).toEqual({
+        country: "EEA",
+        location: "EEA",
+      });
+    });
+
     it("ignores only region headers", async () => {
       const req = createRequest({
         url: "https://privacy.example.com/fides.js",
@@ -53,6 +67,27 @@ describe("getGeolocation", () => {
         },
       });
       const geolocation = await lookupGeolocation(req);
+      expect(geolocation).toBeNull();
+    });
+
+    it("ignores invalid three-character country geolocation headers", async () => {
+      let req = createRequest({
+        url: "https://privacy.example.com/fides.js",
+        headers: {
+          "CloudFront-Viewer-Country": "USA",
+        },
+      });
+      let geolocation = await lookupGeolocation(req);
+      expect(geolocation).toBeNull();
+
+      // Test again including a (seemingly valid!) region
+      req = createRequest({
+        url: "https://privacy.example.com/fides.js",
+        headers: {
+          "CloudFront-Viewer-Country": "USA-NY",
+        },
+      });
+      geolocation = await lookupGeolocation(req);
       expect(geolocation).toBeNull();
     });
 
@@ -86,7 +121,6 @@ describe("getGeolocation", () => {
       const tests = [
         { input: { country: "US", region: undefined }, expected: "US" },
         { input: { country: "us", region: undefined }, expected: "us" },
-        { input: { country: "USA", region: "NY" }, expected: "USA-NY" },
         { input: { country: "SE", region: "O" }, expected: "SE-O" },
         { input: { country: "gb", region: "eng" }, expected: "gb-eng" },
         { input: { country: "RU", region: "PRI" }, expected: "RU-PRI" },
@@ -129,11 +163,37 @@ describe("getGeolocation", () => {
       });
     });
 
+    it("supports the special 'EEA' code in geolocation query param", async () => {
+      const req = createRequest({
+        url: "https://privacy.example.com/fides.js?geolocation=EEA",
+      });
+      const geolocation = await lookupGeolocation(req);
+      expect(geolocation).toEqual({
+        country: "EEA",
+        location: "EEA",
+      });
+    });
+
     it("ignores invalid geolocation query param", async () => {
       const req = createRequest({
         url: "https://privacy.example.com/fides.js?geolocation=America",
       });
       const geolocation = await lookupGeolocation(req);
+      expect(geolocation).toBeNull();
+    });
+
+    it("ignores invalid three-character country codes in geolocation query param", async () => {
+      let req = createRequest({
+        url: "https://privacy.example.com/fides.js?geolocation=USA",
+      });
+      let geolocation = await lookupGeolocation(req);
+      expect(geolocation).toBeNull();
+
+      // Test again including a (seemingly valid!) region
+      req = createRequest({
+        url: "https://privacy.example.com/fides.js?geolocation=USA-NY",
+      });
+      geolocation = await lookupGeolocation(req);
       expect(geolocation).toBeNull();
     });
 
@@ -170,10 +230,6 @@ describe("getGeolocation", () => {
         {
           input: "us",
           expected: { location: "us", country: "us", region: undefined },
-        },
-        {
-          input: "USA-NY",
-          expected: { location: "USA-NY", country: "USA", region: "NY" },
         },
         {
           input: "SE-O",